Move wrapper functions to temp_sensor. scope and inheretence is now unilateral: temp_sensor -> ds18b20 -> onewire
